Rename `range` package to `bound`
