+++ model/src/main/kotlin/io/spine/validation/ValidationPlugin.kt
@@ -30,12 +30,12 @@ import io.spine.protodata.plugin.Plugin
 import io.spine.protodata.plugin.Policy
 import io.spine.protodata.plugin.ViewRepository
 import io.spine.protodata.render.Renderer
-import io.spine.validation.range.MaxFieldView
-import io.spine.validation.range.MaxPolicy
-import io.spine.validation.range.MinFieldView
-import io.spine.validation.range.MinPolicy
-import io.spine.validation.range.RangeFieldView
-import io.spine.validation.range.RangePolicy
+import io.spine.validation.bound.MaxFieldView
+import io.spine.validation.bound.MaxPolicy
+import io.spine.validation.bound.MinFieldView
+import io.spine.validation.bound.MinPolicy
+import io.spine.validation.bound.RangeFieldView
+import io.spine.validation.bound.RangePolicy
 import io.spine.validation.required.RequiredFieldView
 import io.spine.validation.required.RequiredIdOptionPolicy
 import io.spine.validation.required.RequiredIdPatternPolicy
