+++ model/src/main/kotlin/io/spine/validation/bound/MaxPolicy.kt
@@ -24,11 +24,11 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.validation.range
+package io.spine.validation.bound
 
 import io.spine.core.External
 import io.spine.core.Where
-import io.spine.option.MinOption
+import io.spine.option.MaxOption
 import io.spine.protodata.ast.event.FieldOptionDiscovered
 import io.spine.protodata.ast.ref
 import io.spine.protodata.ast.unpack
@@ -36,37 +36,37 @@ import io.spine.protodata.plugin.Policy
 import io.spine.server.event.Just
 import io.spine.server.event.React
 import io.spine.server.event.just
-import io.spine.validation.MIN
+import io.spine.validation.MAX
 import io.spine.validation.OPTION_NAME
 import io.spine.validation.defaultMessage
-import io.spine.validation.event.MinFieldDiscovered
-import io.spine.validation.event.minFieldDiscovered
+import io.spine.validation.event.MaxFieldDiscovered
+import io.spine.validation.event.maxFieldDiscovered
 
 /**
- * A policy to add a validation rule to a type whenever the `(min)` field option
+ * A policy to add a validation rule to a type whenever the `(max)` field option
  * is discovered.
  */
-internal class MinPolicy : Policy<FieldOptionDiscovered>() {
+internal class MaxPolicy : Policy<FieldOptionDiscovered>() {
 
     @React
     override fun whenever(
-        @External @Where(field = OPTION_NAME, equals = MIN)
+        @External @Where(field = OPTION_NAME, equals = MAX)
         event: FieldOptionDiscovered
-    ): Just<MinFieldDiscovered> {
+    ): Just<MaxFieldDiscovered> {
         val field = event.subject
         val file = event.file
-        val primitiveType = checkFieldType(field, file, MIN)
+        val primitiveType = checkFieldType(field, file, MAX)
 
-        val option = event.option.unpack<MinOption>()
-        val context = BoundContext(MIN, primitiveType, field, file)
+        val option = event.option.unpack<MaxOption>()
+        val context = BoundContext(MAX, primitiveType, field, file)
         val kotlinBound = context.checkNumericBound(option.value, option.exclusive)
 
         val message = option.errorMsg.ifEmpty { option.descriptorForType.defaultMessage }
-        return minFieldDiscovered {
+        return maxFieldDiscovered {
             id = field.ref
             subject = field
             errorMessage = message
-            this.min = option.value
+            this.max = option.value
             bound = kotlinBound.toProto()
             this.file = file
         }.just()
