+++ model/src/main/kotlin/io/spine/validation/bound/MinFieldView.kt
@@ -24,35 +24,26 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.validation.range
+package io.spine.validation.bound
 
-import io.spine.protodata.ast.Field
-import io.spine.protodata.ast.File
-import io.spine.protodata.ast.PrimitiveType
-import io.spine.validation.RANGE
-
-/**
- * The context of validating a numeric option that constrains a field's value
- * with a minimum or maximum bound.
- *
- * Contains the data required to report a compilation error for the option.
- */
-internal open class BoundContext(
-    val optionName: String,
-    val primitiveType: PrimitiveType,
-    val field: Field,
-    val file: File
-)
+import io.spine.core.Subscribe
+import io.spine.protodata.ast.FieldRef
+import io.spine.protodata.plugin.View
+import io.spine.server.entity.alter
+import io.spine.validation.MinField
+import io.spine.validation.event.MinFieldDiscovered
 
 /**
- * The [BoundContext] for the `(range)` option.
- *
- * Introduces the [range] property to report the originally passed range definition
- * in compilation errors.
+ * A view of a field that is marked with the `(min)` option.
  */
-internal class RangeContext(
-    val range: String,
-    primitiveType: PrimitiveType,
-    field: Field,
-    file: File
-) : BoundContext(RANGE, primitiveType, field, file)
+internal class MinFieldView : View<FieldRef, MinField, MinField.Builder>() {
+
+    @Subscribe
+    fun on(e: MinFieldDiscovered) = alter {
+        subject = e.subject
+        errorMessage = e.errorMessage
+        min = e.min
+        bound = e.bound
+        file = e.file
+    }
+}
